+++ CHANGELOG.rst
@@ -6,7 +6,9 @@ Changelog
 Current
 -------
 
-- Fix `@api.expect(..., validate=False)` decorators for an :class:`Api` where `validate=True` is set on the constructor
+- Add new `Wildcard` fields (:pr:`255`)
+- Fix ABC deprecation warnings (:pr:`580`)
+- Fix `@api.expect(..., validate=False)` decorators for an :class:`Api` where `validate=True` is set on the constructor (:issue:`609`, :pr:`610`)
 - Ensure `basePath` is always a path
 
 0.12.1 (2018-09-28)